+++ ios/versioned-react-native/ABI39_0_0/Expo/EXUpdates/ABI39_0_0EXUpdates/Update/ABI39_0_0EXUpdatesLegacyUpdate.m
@@ -108,7 +108,7 @@ static NSString * const ABI39_0_0EXUpdatesExpoTestDomain = @"expo.test";
     [processedAssets addObject:asset];
   }
 
-  update.metadata = manifest;
+  update.manifest = manifest;
   update.keep = YES;
   update.bundleUrl = bundleUrl;
   update.assets = processedAssets;