+++ packages/expo-updates/ios/EXUpdates/Update/EXUpdatesNewUpdate.m
@@ -83,7 +83,7 @@ NS_ASSUME_NONNULL_BEGIN
   update.keep = YES;
   update.bundleUrl = bundleUrl;
   update.assets = processedAssets;
-  update.metadata = manifest.rawManifestJSON;
+  update.manifest = manifest.rawManifestJSON;
   
   if ([response isKindOfClass:[NSHTTPURLResponse class]]) {
     NSDictionary *headersDictionary = ((NSHTTPURLResponse *)response).allHeaderFields;